Fix printf/scanf-style format strings and cast arguments where needed.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@1487 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/contrib/bogogrep.c
@@ -90,7 +90,7 @@ int main(int argc, char **argv) {
 		int l = strcspn((const char *)i, "\n");
 
 		if (l > base - i + st.st_size) l = base - i + st.st_size;
-		printf("%d:", i - base);
+		printf("%ld:", (long)(i - base));
 		fwrite(i, 1, strcspn((const char *)i, "\n"), stdout);
 		if (EOF == puts("")) {
 		    perror("stdout");
